Fix token refresh issue
Fixed a bug where if the user access token was expired, the user will be redirected to `/auth/refresh` to get a new access token. Here if the access token was fetch or not fetch, the user will always be redirected to `/` and here if the app found that there is no access token, it will again redirect back to `/auth/refresh` to get a new access token. This will cause an infinite loop of redirects. Now, if the access token is not fetched, the user will be logged out and redirected to `/`.